+++ cmd/snap-update-ns/export_test.go
@@ -20,12 +20,7 @@
 package main
 
 import (
-	"os"
-	"syscall"
-
 	. "gopkg.in/check.v1"
-
-	"github.com/snapcore/snapd/osutil/sys"
 )
 
 var (
@@ -38,101 +33,15 @@ var (
 	// utils
 	PlanWritableMimic = planWritableMimic
 	ExecWritableMimic = execWritableMimic
-
 	// main
 	ComputeAndSaveChanges = computeAndSaveChanges
 	ApplyUserFstab        = applyUserFstab
-
 	// bootstrap
 	ClearBootstrapError = clearBootstrapError
-
 	// trespassing
 	IsReadOnly                   = isReadOnly
 	IsPrivateTmpfsCreatedBySnapd = isPrivateTmpfsCreatedBySnapd
 )
-
-// SystemCalls encapsulates various system interactions performed by this module.
-type SystemCalls interface {
-	OsLstat(name string) (os.FileInfo, error)
-	SysLstat(name string, buf *syscall.Stat_t) error
-	ReadDir(dirname string) ([]os.FileInfo, error)
-	Symlinkat(oldname string, dirfd int, newname string) error
-	Readlinkat(dirfd int, path string, buf []byte) (int, error)
-	Remove(name string) error
-
-	Close(fd int) error
-	Fchdir(fd int) error
-	Fchown(fd int, uid sys.UserID, gid sys.GroupID) error
-	Mkdirat(dirfd int, path string, mode uint32) error
-	Mount(source string, target string, fstype string, flags uintptr, data string) (err error)
-	Open(path string, flags int, mode uint32) (fd int, err error)
-	Openat(dirfd int, path string, flags int, mode uint32) (fd int, err error)
-	Unmount(target string, flags int) error
-	Fstat(fd int, buf *syscall.Stat_t) error
-	Fstatfs(fd int, buf *syscall.Statfs_t) error
-}
-
-// MockSystemCalls replaces real system calls with those of the argument.
-func MockSystemCalls(sc SystemCalls) (restore func()) {
-	// save
-	oldOsLstat := osLstat
-	oldRemove := osRemove
-	oldIoutilReadDir := ioutilReadDir
-
-	oldSysClose := sysClose
-	oldSysFchown := sysFchown
-	oldSysMkdirat := sysMkdirat
-	oldSysMount := sysMount
-	oldSysOpen := sysOpen
-	oldSysOpenat := sysOpenat
-	oldSysUnmount := sysUnmount
-	oldSysSymlinkat := sysSymlinkat
-	oldReadlinkat := sysReadlinkat
-	oldFstat := sysFstat
-	oldFstatfs := sysFstatfs
-	oldSysFchdir := sysFchdir
-	oldSysLstat := sysLstat
-
-	// override
-	osLstat = sc.OsLstat
-	osRemove = sc.Remove
-	ioutilReadDir = sc.ReadDir
-
-	sysClose = sc.Close
-	sysFchown = sc.Fchown
-	sysMkdirat = sc.Mkdirat
-	sysMount = sc.Mount
-	sysOpen = sc.Open
-	sysOpenat = sc.Openat
-	sysUnmount = sc.Unmount
-	sysSymlinkat = sc.Symlinkat
-	sysReadlinkat = sc.Readlinkat
-	sysFstat = sc.Fstat
-	sysFstatfs = sc.Fstatfs
-	sysFchdir = sc.Fchdir
-	sysLstat = sc.SysLstat
-
-	return func() {
-		// restore
-		osLstat = oldOsLstat
-		osRemove = oldRemove
-		ioutilReadDir = oldIoutilReadDir
-
-		sysClose = oldSysClose
-		sysFchown = oldSysFchown
-		sysMkdirat = oldSysMkdirat
-		sysMount = oldSysMount
-		sysOpen = oldSysOpen
-		sysOpenat = oldSysOpenat
-		sysUnmount = oldSysUnmount
-		sysSymlinkat = oldSysSymlinkat
-		sysReadlinkat = oldReadlinkat
-		sysFstat = oldFstat
-		sysFstatfs = oldFstatfs
-		sysFchdir = oldSysFchdir
-		sysLstat = oldSysLstat
-	}
-}
 
 func MockFreezerCgroupDir(c *C) (restore func()) {
 	old := freezerCgroupDir
@@ -154,30 +63,10 @@ func MockChangePerform(f func(chg *Change, as *Assumptions) ([]*Change, error)) 
 	}
 }
 
-func MockReadDir(fn func(string) ([]os.FileInfo, error)) (restore func()) {
-	old := ioutilReadDir
-	ioutilReadDir = fn
-	return func() {
-		ioutilReadDir = old
-	}
-}
-
-func MockReadlink(fn func(string) (string, error)) (restore func()) {
-	old := osReadlink
-	osReadlink = fn
-	return func() {
-		osReadlink = old
-	}
-}
-
 func (as *Assumptions) IsRestricted(path string) bool {
 	return as.isRestricted(path)
 }
 
 func (as *Assumptions) PastChanges() []*Change {
 	return as.pastChanges
-}
-
-func (as *Assumptions) CanWriteToDirectory(dirFd int, dirName string) (bool, error) {
-	return as.canWriteToDirectory(dirFd, dirName)
 }
